Refactor TestUtil.java and IbanServiceTest.java
Refactor the TestUtil.java by adding two new methods `fileToObject(String relativePath, Class<T> tClass)` and `fileToObject(String relativePath, TypeReference<T> typeReference)`. These methods read a JSON file and convert it into an object of the given type.

In IbanServiceTest.java, add `@AutoConfigureMockMvc` and `@WithMockUser(username = "user1", password = "pwd", roles = "USER")` annotations. This configures the MockMvc and sets a mock user for testing.

Add three new test methods in IbanServiceTest.java: `getIban()`, `getCreditorInstitutionIbans()`, and `createCreditorInstitutionIbans()`. These methods test different functionalities of the `IbanService` class.

In the test method `exportIbansToCsv()`, test the functionality of exporting IBANs to a CSV file and assert that the result is not null.

In the test method `getIban()`, mock the `getCreditorInstitutionIbans()` method of the `apiConfigClient` and return a mocked response. Then, call the `getIban()` method of the `ibanService` and assert that the response is not null and contains a list of IBANs with a size of 28.

In the test method `getCreditorInstitutionIbans()`, use the `MvcMock` to perform a GET request to the `/creditor-institutions/11111/ibans` endpoint, asserting that the status is OK and the content type is JSON.

In the test method `createCreditorInstitutionIbans()`, use the `MvcMock` to perform a POST request to the `/creditor-institutions/11111/ibans` endpoint with a JSON payload, asserting that the status is Created and the content type is JSON.

Ignore any changes to imports and requires.